+++ src/string.h
@@ -18,10 +18,11 @@
 #ifndef STRINGS_H
 #define STRINGS_H
 
-#include "types.h"
+#include <stdint.h>
+#include <stddef.h>
 
-char* uitoa(size_t value, char* result, int base);
-size_t strlen(char *str);
+char* uitoa(uint32_t value, char* result, int base);
+uint32_t strlen(char *str);
 void printk(const char *fmt, ...);
 
 #endif
